fuzzy_search/rt_text_search/requirements.txt
python-Levenshtein~=0.12
rapidfuzz~=1.2
fuzzywuzzy~=0.18
numpy~=1.20
pandas~=1.2